bug fix: always force knative and kubernetes layers to exist
The cluster codegen assumes that the layer for its assigned
nature will be generated, so ensure that is true.  Considered
an alternative fix of making inclusion of the nature layer
conditional on it existing, but that decided having an empty
layer gave more flexibility for post solsa-yaml manipulation
of the generated yaml code.