simpler ROKS Route specification
On ROKS, if the host field of a Route is defined to be '', then
when the route is actually created it gets a reasonable default
vhost of <name-of-Route>-namespace.cluster-dns-subdomain.  Exploiting
this avoids the need to actually know the subdomain when writing the
solsa.yaml stanza for the cluster, thus simplifying the user experience.